+++ src/code42cli/cmds/alerts.py
@@ -320,7 +320,9 @@ UPDATE_ALERT_CSV_HEADERS = ["id", "state", "note"]
 
 
 update_alerts_generate_template = generate_template_cmd_factory(
-    group_name="alerts", commands_dict={"update": UPDATE_ALERT_CSV_HEADERS},
+    group_name="alerts",
+    commands_dict={"update": UPDATE_ALERT_CSV_HEADERS},
+    help_message="Generate the CSV template needed for bulk updating alerts.",
 )
 bulk.add_command(update_alerts_generate_template)