+++ CHANGELOG.md
@@ -8,6 +8,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## Unreleased
+
+### Fixed
+
+- Bug where some CSV outputs on Windows would have an extra newline between the rows.
+
+### Changed
+
+- `code42 alert-rules list` now uses outputs via pager when the results are greater than 10.
+
+- `code42 cases list` now uses outputs via pager when the results are greater than 10.
+
 ## 1.4.1 - 2021-04-15
 
 ### Added