+++ CHANGELOG.md
@@ -16,9 +16,9 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 ### Changed
 
-- `code42 alert-rules list` now uses outputs via pager when the results are greater than 10.
+- `code42 alert-rules list` now outputs via a pager when more than 10 rules.
 
-- `code42 cases list` now uses outputs via pager when the results are greater than 10.
+- `code42 cases list` now outputs via a pager when more than 10 cases.
 
 ## 1.4.1 - 2021-04-15