+++ CHANGELOG.md
@@ -21,7 +21,7 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 
 - New command `code42 alerts update` that can update an alert's state or note.
 
-- New command `code42 alerts bulk generate-tempate` for generating CSV templates for bulk 
+- New command `code42 alerts bulk generate-tempate` for generating CSV templates for bulk
   commands.
 
 - New command `code42 alerts bulk update` for bulk update alerts.